Move helpers into kiosks-content
